requirements.txt
# Django & REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database & ORM
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Server & Static Files
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
Pillow==10.0.1
requests==2.31.0
